Admin - 상품 관리 페이지에 상품 삭제 추가

+++ src/main/java/com/foot/dto/products/ProductResponseDto.java
@@ -12,7 +12,7 @@ public class ProductResponseDto {
     Long price;
     String ModelPicture;
     double discountRate;
-    double discountPrice;
+    Long discountPrice;
 //    String brand;
 
     public ProductResponseDto(Product product){
